Extract and adjust relative URL handling
The code for generating relative URLs is extracted into a public method,
making it easier to reuse. In addition, the last component of the
current URL is now treated as a file instead of a directory, meaning
that linking to ../foo.css from /foo/bar/baz results in /foo/foo.css,
instead of /foo/bar/foo.css.

Changelog: changed